# ARROW-11567: [C++][Compute] Improve variance kernel precision