Variance kernel suffers from round-off errors due to naive summation.
This patch implements pairwise summation algorithm to improve variance
kernel precision.

Pairwise summation also results to better performance for normal cases.
- ~1x improvement for 0.01%, 0% null count
- moderate improvement for 1% null count
- ~0.5x drop for 10%, 50% null count